requirements.txt
asgiref-3.5.2 
django-4.1 
requests-2.28.1
sqlparse-0.4.2 
tzdata-2022.2